[#175798769] Splitted integration test suite to differentiate stub PM / dockerized PM
